+++ src/shared/components/case-editor/case-edit/case-edit.component.spec.ts
@@ -18,7 +18,7 @@ import { WizardPageField } from '../domain/wizard-page-field.model';
 import { CaseField } from '../../../domain/definition/case-field.model';
 import { Wizard } from '../domain/wizard.model';
 import { WizardPage } from '../domain/wizard-page.model';
-import { Profile } from '../../../domain';
+import { FieldType, Profile } from '../../../domain';
 import createSpyObj = jasmine.createSpyObj;
 
 describe('CaseEditComponent', () => {
@@ -44,26 +44,13 @@ describe('CaseEditComponent', () => {
     ]
   );
 
-  const WIZARD_PAGE_FIELD_WITH_SHOW_CONDITION: WizardPageField = {
-    case_field_id: 'PersonFirstName'
-  };
-
-  const WIZARD_PAGE_1: WizardPageField = {
-    case_field_id: 'PersonFirstName'
-  };
-
-  const WIZARD_PAGE_2: WizardPageField = {
-    case_field_id: 'PersonLastName'
-  };
-
-  const WIZARD_PAGE_3: WizardPageField = {
-    case_field_id: 'Address'
-  };
-
   const CASE_FIELD_WITH_SHOW_CONDITION: CaseField = <CaseField>({
     id: 'PersonFirstName',
     label: 'First name',
-    field_type: null,
+    field_type: {
+      id: 'Text',
+      type: 'Text'
+    },
     display_context: 'READONLY',
     show_condition: 'PersonLastName=\"Smith\"'
   });
@@ -71,23 +58,140 @@ describe('CaseEditComponent', () => {
   const CASE_FIELD_1: CaseField = <CaseField>({
     id: 'PersonFirstName',
     label: 'First name',
-    field_type: null,
+    field_type: {
+      id: 'Text',
+      type: 'Text'
+    },
     display_context: 'READONLY'
   });
 
   const CASE_FIELD_2: CaseField = <CaseField>({
     id: 'PersonLastName',
-    label: 'First name',
-    field_type: null,
+    label: 'Last name',
+    field_type: {
+      id: 'Text',
+      type: 'Text'
+    },
     display_context: 'READONLY'
+  });
+
+  const PERSON_NAME_FIELD_TYPE: FieldType = {
+    id: 'PersonName',
+    type: 'Complex',
+    complex_fields: [
+      {
+        id: 'PersonMiddleName',
+        label: 'Middle name',
+        field_type: {
+          id: 'Text',
+          type: 'Text'
+        },
+        display_context: 'READONLY'
+      } as CaseField
+    ]
+  }
+
+  const CASE_FIELD_2_COMPLEX: CaseField = <CaseField>({
+    id: 'PersonLastName',
+    label: 'Last name',
+    field_type: PERSON_NAME_FIELD_TYPE,
+    display_context: 'READONLY'
+  });
+
+  const CASE_FIELD_2_COLLECTION: CaseField = <CaseField>({
+    id: 'PersonLastNameCollection',
+    label: 'Last name (collection)',
+    field_type: {
+      id: 'LastNameCollection',
+      type: 'Collection',
+      collection_field_type: PERSON_NAME_FIELD_TYPE
+    },
+    display_context: 'READONLY',
+    value: [
+      {
+        id: '0',
+        value: {
+          PersonMiddleName: 'John'
+        }
+      }
+    ]
   });
 
   const CASE_FIELD_3: CaseField = <CaseField>({
     id: 'Address',
     label: 'Address',
-    field_type: null,
+    field_type: {
+      id: 'Text',
+      type: 'Text'
+    },
     display_context: 'READONLY'
   });
+
+  const ADDRESS_DETAILS_FIELD_TYPE: FieldType = {
+    id: 'AddressDetails',
+    type: 'Complex',
+    complex_fields: [
+      {
+        id: 'AddressLine1',
+        label: 'Address line 1',
+        field_type: {
+          id: 'Text',
+          type: 'Text'
+        },
+        display_context: 'READONLY'
+      } as CaseField
+    ]
+  }
+
+  const CASE_FIELD_3_COMPLEX: CaseField = <CaseField>({
+    id: 'Address',
+    label: 'Address',
+    field_type: ADDRESS_DETAILS_FIELD_TYPE,
+    display_context: 'READONLY'
+  });
+
+  const CASE_FIELD_3_COLLECTION: CaseField = <CaseField>({
+    id: 'AddressCollection',
+    label: 'Address (collection)',
+    field_type: {
+      id: 'AddressDetailsCollection',
+      type: 'Collection',
+      collection_field_type: ADDRESS_DETAILS_FIELD_TYPE
+    },
+    display_context: 'READONLY',
+    value: [
+      {
+        id: '0',
+        value: {
+          AddressLine1: 'Street'
+        }
+      }
+    ]
+  });
+
+  const WIZARD_PAGE_FIELD_WITH_SHOW_CONDITION: WizardPageField = {
+    case_field_id: CASE_FIELD_WITH_SHOW_CONDITION.id
+  };
+
+  const WIZARD_PAGE_1: WizardPageField = {
+    case_field_id: CASE_FIELD_1.id
+  };
+
+  const WIZARD_PAGE_2: WizardPageField = {
+    case_field_id: CASE_FIELD_2.id
+  };
+
+  const WIZARD_PAGE_2_COLLECTION: WizardPageField = {
+    case_field_id: CASE_FIELD_2_COLLECTION.id
+  };
+
+  const WIZARD_PAGE_3: WizardPageField = {
+    case_field_id: CASE_FIELD_3.id
+  };
+
+  const WIZARD_PAGE_3_COLLECTION: WizardPageField = {
+    case_field_id: CASE_FIELD_3_COLLECTION.id
+  };
 
   let fixture: ComponentFixture<CaseEditComponent>;
   let component: CaseEditComponent;
@@ -278,7 +382,7 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.nextPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
+            data: new FormGroup({
               PersonFirstName: new FormControl('John'),
               PersonLastName: new FormControl('Smith')
             })
@@ -301,10 +405,10 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.nextPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('John'),
-                PersonLastName: new FormControl('Smith')
-              })
+            data: new FormGroup({
+              PersonFirstName: new FormControl('John'),
+              PersonLastName: new FormControl('Smith')
+            })
           });
           fixture.detectChanges();
 
@@ -324,11 +428,11 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.nextPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('John'),
-                PersonLastName: new FormControl('Other')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('John'),
+              PersonLastName: new FormControl('Other')
+            })
+          });
           fixture.detectChanges();
 
           component.next('somePage');
@@ -347,11 +451,11 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.nextPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormGroup({PersonMiddleName: new FormControl('John')}),
-                PersonLastName: new FormControl('Other')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormGroup({PersonMiddleName: new FormControl('John')}),
+              PersonLastName: new FormControl('Other')
+            })
+          });
           fixture.detectChanges();
 
           component.next('somePage');
@@ -370,11 +474,11 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.nextPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormArray([new FormGroup({PersonMiddleName: new FormControl('John')})]),
-                PersonLastName: new FormControl('Other')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormArray([new FormGroup({PersonMiddleName: new FormControl('John')})]),
+              PersonLastName: new FormControl('Other')
+            })
+          });
           fixture.detectChanges();
 
           component.next('somePage');
@@ -395,7 +499,7 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.previousPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
+            data: new FormGroup({
               PersonFirstName: new FormControl('John'),
               PersonLastName: new FormControl('Smith')
             })
@@ -418,11 +522,11 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.previousPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('John'),
-                PersonLastName: new FormControl('Smith')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('John'),
+              PersonLastName: new FormControl('Smith')
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -441,11 +545,11 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.previousPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('John'),
-                PersonLastName: new FormControl('Other')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('John'),
+              PersonLastName: new FormControl('Other')
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -464,11 +568,11 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.previousPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormGroup({PersonMiddleName: new FormControl('John')}),
-                PersonLastName: new FormControl('Other')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormGroup({PersonMiddleName: new FormControl('John')}),
+              PersonLastName: new FormControl('Other')
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -487,11 +591,11 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           wizard.previousPage.and.returnValue(new WizardPage());
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormArray([new FormGroup({PersonMiddleName: new FormControl('John')})]),
-                PersonLastName: new FormControl('Other')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormArray([new FormGroup({PersonMiddleName: new FormControl('John')})]),
+              PersonLastName: new FormControl('Other')
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -516,17 +620,17 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let nextPage = new WizardPage();
           nextPage.show_condition = 'PersonFirstName=\"John\"';
-          nextPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
+          nextPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3];
           nextPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
           wizard.pages = [currentPage, nextPage];
           wizard.nextPage.and.returnValue(nextPage);
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('John'),
-                PersonLastName: new FormControl('Smith'),
-                Address: new FormControl('Some street')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('John'),
+              PersonLastName: new FormControl('Smith'),
+              Address: new FormControl('Some street')
+            })
+          });
           fixture.detectChanges();
 
           component.next('somePage');
@@ -538,7 +642,7 @@ describe('CaseEditComponent', () => {
           expect(component.form.get('data').get(CASE_FIELD_3.id)).not.toBeNull();
         });
 
-        it('should navigate to next page when next is called and clear hidden simple form field', () => {
+        it('should navigate to next page when next is called and retain hidden simple form fields with retain_hidden_value = true', () => {
           component.wizard = wizard;
           let currentPage = new WizardPage();
           currentPage.wizard_page_fields = [WIZARD_PAGE_1];
@@ -546,17 +650,18 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let nextPage = new WizardPage();
           nextPage.show_condition = 'PersonFirstName=\"John\"';
-          nextPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
+          CASE_FIELD_3.retain_hidden_value = true;
+          nextPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3];
           nextPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
           wizard.nextPage.and.returnValue(nextPage);
           wizard.pages = [currentPage, nextPage];
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('Other'),
-                PersonLastName: new FormControl('Smith'),
-                Address: new FormControl('Some street')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('Other'),
+              PersonLastName: new FormControl('Smith'),
+              Address: new FormControl('Some street')
+            })
+          });
           fixture.detectChanges();
 
           component.next('somePage');
@@ -564,8 +669,8 @@ describe('CaseEditComponent', () => {
           expect(wizard.nextPage).toHaveBeenCalled();
           expect(routerStub.navigate).toHaveBeenCalled();
           expect(component.form.get('data').get(CASE_FIELD_1.id)).not.toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_2.id)).toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_3.id)).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_2.id).value).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_3.id).value).not.toBeNull();
         });
 
         it('should navigate to next page when next is called and clear hidden complex form field', () => {
@@ -576,17 +681,19 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let nextPage = new WizardPage();
           nextPage.show_condition = 'PersonFirstName=\"John\"';
-          nextPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
+          CASE_FIELD_2_COMPLEX.retain_hidden_value = true;
+          CASE_FIELD_3_COMPLEX.retain_hidden_value = false;
+          nextPage.case_fields = [CASE_FIELD_2_COMPLEX, CASE_FIELD_3_COMPLEX];
           nextPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
           wizard.nextPage.and.returnValue(nextPage);
           wizard.pages = [currentPage, nextPage];
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('Other'),
-                PersonLastName: new FormGroup({PersonMiddleName: new FormControl('John')}),
-                Address: new FormGroup({AddressLine1: new FormControl('Street')})
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('Other'),
+              PersonLastName: new FormGroup({PersonMiddleName: new FormControl('John')}),
+              Address: new FormGroup({AddressLine1: new FormControl('Street')})
+            })
+          });
           fixture.detectChanges();
 
           component.next('somePage');
@@ -594,8 +701,14 @@ describe('CaseEditComponent', () => {
           expect(wizard.nextPage).toHaveBeenCalled();
           expect(routerStub.navigate).toHaveBeenCalled();
           expect(component.form.get('data').get(CASE_FIELD_1.id)).not.toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_2.id)).toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_3.id)).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_2_COMPLEX.id)).not.toBeNull();
+          expect(component.form.get('data').get(`${CASE_FIELD_2_COMPLEX.id}.PersonMiddleName`)).not.toBeNull();
+          // 'PersonMiddleName' value expected to be null because this sub-field does not have
+          // retain_hidden_value = true, even though its parent Complex field does
+          expect(component.form.get('data').get(`${CASE_FIELD_2_COMPLEX.id}.PersonMiddleName`).value).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_3_COMPLEX.id)).not.toBeNull();
+          expect(component.form.get('data').get(`${CASE_FIELD_3_COMPLEX.id}.AddressLine1`)).not.toBeNull();
+          expect(component.form.get('data').get(`${CASE_FIELD_3_COMPLEX.id}.AddressLine1`).value).toBeNull();
         });
 
         it('should navigate to next page when next is called and clear hidden collection form field', () => {
@@ -606,17 +719,23 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let nextPage = new WizardPage();
           nextPage.show_condition = 'PersonFirstName=\"John\"';
-          nextPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
-          nextPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
+          CASE_FIELD_2_COLLECTION.retain_hidden_value = true;
+          CASE_FIELD_3_COLLECTION.retain_hidden_value = false;
+          nextPage.case_fields = [CASE_FIELD_2_COLLECTION, CASE_FIELD_3_COLLECTION];
+          nextPage.wizard_page_fields = [WIZARD_PAGE_2_COLLECTION, WIZARD_PAGE_3_COLLECTION];
           wizard.nextPage.and.returnValue(nextPage);
           wizard.pages = [currentPage, nextPage];
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('Other'),
-                PersonLastName: new FormArray([new FormGroup({PersonMiddleName: new FormControl('John')})]),
-                Address: new FormArray([new FormGroup({AddressLine1: new FormControl('Street')})])
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('Other'),
+              PersonLastNameCollection: new FormArray([new FormGroup({
+                value: new FormGroup({PersonMiddleName: new FormControl('John')})
+              })]),
+              AddressCollection: new FormArray([new FormGroup({
+                value: new FormGroup({AddressLine1: new FormControl('Street')})
+              })])
+            })
+          });
           fixture.detectChanges();
 
           component.next('somePage');
@@ -624,8 +743,18 @@ describe('CaseEditComponent', () => {
           expect(wizard.nextPage).toHaveBeenCalled();
           expect(routerStub.navigate).toHaveBeenCalled();
           expect(component.form.get('data').get(CASE_FIELD_1.id)).not.toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_2.id)).toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_3.id)).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_2_COLLECTION.id)).not.toBeNull();
+          expect((component.form.get('data').get(CASE_FIELD_2_COLLECTION.id) as FormArray).at(0)
+            .get('value.PersonMiddleName')).not.toBeNull();
+          // 'PersonMiddleName' value expected to be null because this sub-field does not have
+          // retain_hidden_value = true, even though its top-level collection field does
+          expect((component.form.get('data').get(CASE_FIELD_2_COLLECTION.id) as FormArray).at(0)
+            .get('value.PersonMiddleName').value).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_3_COLLECTION.id)).not.toBeNull();
+          expect((component.form.get('data').get(CASE_FIELD_3_COLLECTION.id) as FormArray).at(0)
+            .get('value.AddressLine1')).not.toBeNull();
+          expect((component.form.get('data').get(CASE_FIELD_3_COLLECTION.id) as FormArray).at(0)
+            .get('value.AddressLine1').value).toBeNull();
         });
       });
 
@@ -639,17 +768,17 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let previousPage = new WizardPage();
           previousPage.show_condition = 'PersonFirstName=\"John\"';
-          previousPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
+          previousPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3];
           previousPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
           wizard.pages = [previousPage, currentPage];
           wizard.previousPage.and.returnValue(previousPage);
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('John'),
-                PersonLastName: new FormControl('Smith'),
-                Address: new FormControl('Some street')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('John'),
+              PersonLastName: new FormControl('Smith'),
+              Address: new FormControl('Some street')
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -661,7 +790,8 @@ describe('CaseEditComponent', () => {
           expect(component.form.get('data').get(CASE_FIELD_3.id)).not.toBeNull();
         });
 
-        it('should navigate to previous page when previous is called and clear hidden simple form field', () => {
+        it('should navigate to previous page when previous is called and retain hidden simple form fields with ' +
+        'retain_hidden_value = true', () => {
           component.wizard = wizard;
           let currentPage = new WizardPage();
           currentPage.wizard_page_fields = [WIZARD_PAGE_1];
@@ -669,17 +799,18 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let previousPage = new WizardPage();
           previousPage.show_condition = 'PersonFirstName=\"John\"';
-          previousPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
+          CASE_FIELD_3.retain_hidden_value = true;
+          previousPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3];
           previousPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
           wizard.previousPage.and.returnValue(previousPage);
           wizard.pages = [previousPage, currentPage];
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('Other'),
-                PersonLastName: new FormControl('Smith'),
-                Address: new FormControl('Some street')
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('Other'),
+              PersonLastName: new FormControl('Smith'),
+              Address: new FormControl('Some street')
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -687,8 +818,8 @@ describe('CaseEditComponent', () => {
           expect(wizard.previousPage).toHaveBeenCalled();
           expect(routerStub.navigate).toHaveBeenCalled();
           expect(component.form.get('data').get(CASE_FIELD_1.id)).not.toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_2.id)).toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_3.id)).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_2.id).value).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_3.id).value).not.toBeNull();
         });
 
         it('should navigate to previous page when previous is called and clear hidden complex form field', () => {
@@ -699,17 +830,19 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let previousPage = new WizardPage();
           previousPage.show_condition = 'PersonFirstName=\"John\"';
-          previousPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
+          CASE_FIELD_2_COMPLEX.retain_hidden_value = true;
+          CASE_FIELD_3_COMPLEX.retain_hidden_value = false;
+          previousPage.case_fields = [CASE_FIELD_2_COMPLEX, CASE_FIELD_3_COMPLEX];
           previousPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
           wizard.previousPage.and.returnValue(previousPage);
           wizard.pages = [previousPage, currentPage];
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('Other'),
-                PersonLastName: new FormGroup({PersonMiddleName: new FormControl('John')}),
-                Address: new FormGroup({AddressLine1: new FormControl('Street')})
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('Other'),
+              PersonLastName: new FormGroup({PersonMiddleName: new FormControl('John')}),
+              Address: new FormGroup({AddressLine1: new FormControl('Street')})
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -717,8 +850,14 @@ describe('CaseEditComponent', () => {
           expect(wizard.previousPage).toHaveBeenCalled();
           expect(routerStub.navigate).toHaveBeenCalled();
           expect(component.form.get('data').get(CASE_FIELD_1.id)).not.toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_2.id)).toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_3.id)).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_2_COMPLEX.id)).not.toBeNull();
+          expect(component.form.get('data').get(`${CASE_FIELD_2_COMPLEX.id}.PersonMiddleName`)).not.toBeNull();
+          // 'PersonMiddleName' value expected to be null because this sub-field does not have
+          // retain_hidden_value = true, even though its parent Complex field does
+          expect(component.form.get('data').get(`${CASE_FIELD_2_COMPLEX.id}.PersonMiddleName`).value).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_3_COMPLEX.id)).not.toBeNull();
+          expect(component.form.get('data').get(`${CASE_FIELD_3_COMPLEX.id}.AddressLine1`)).not.toBeNull();
+          expect(component.form.get('data').get(`${CASE_FIELD_3_COMPLEX.id}.AddressLine1`).value).toBeNull();
         });
 
         it('should navigate to previous page when previous is called and clear hidden collection form field', () => {
@@ -729,17 +868,23 @@ describe('CaseEditComponent', () => {
           wizard.getPage.and.returnValue(currentPage);
           let previousPage = new WizardPage();
           previousPage.show_condition = 'PersonFirstName=\"John\"';
-          previousPage.case_fields = [CASE_FIELD_2, CASE_FIELD_3]
-          previousPage.wizard_page_fields = [WIZARD_PAGE_2, WIZARD_PAGE_3];
+          CASE_FIELD_2_COLLECTION.retain_hidden_value = true;
+          CASE_FIELD_3_COLLECTION.retain_hidden_value = false;
+          previousPage.case_fields = [CASE_FIELD_2_COLLECTION, CASE_FIELD_3_COLLECTION];
+          previousPage.wizard_page_fields = [WIZARD_PAGE_2_COLLECTION, WIZARD_PAGE_3_COLLECTION];
           wizard.previousPage.and.returnValue(previousPage);
           wizard.pages = [currentPage, previousPage];
           component.form = new FormGroup({
-            data : new FormGroup({
-                PersonFirstName: new FormControl('Other'),
-                PersonLastName: new FormArray([new FormGroup({PersonMiddleName: new FormControl('John')})]),
-                Address: new FormArray([new FormGroup({AddressLine1: new FormControl('Street')})])
-              })
-            });
+            data: new FormGroup({
+              PersonFirstName: new FormControl('Other'),
+              PersonLastNameCollection: new FormArray([new FormGroup({
+                value: new FormGroup({PersonMiddleName: new FormControl('John')})
+              })]),
+              AddressCollection: new FormArray([new FormGroup({
+                value: new FormGroup({AddressLine1: new FormControl('Street')})
+              })])
+            })
+          });
           fixture.detectChanges();
 
           component.previous('somePage');
@@ -747,8 +892,18 @@ describe('CaseEditComponent', () => {
           expect(wizard.previousPage).toHaveBeenCalled();
           expect(routerStub.navigate).toHaveBeenCalled();
           expect(component.form.get('data').get(CASE_FIELD_1.id)).not.toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_2.id)).toBeNull();
-          expect(component.form.get('data').get(CASE_FIELD_3.id)).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_2_COLLECTION.id)).not.toBeNull();
+          expect((component.form.get('data').get(CASE_FIELD_2_COLLECTION.id) as FormArray).at(0)
+            .get('value.PersonMiddleName')).not.toBeNull();
+          // 'PersonMiddleName' value expected to be null because this sub-field does not have
+          // retain_hidden_value = true, even though its top-level collection field does
+          expect((component.form.get('data').get(CASE_FIELD_2_COLLECTION.id) as FormArray).at(0)
+            .get('value.PersonMiddleName').value).toBeNull();
+          expect(component.form.get('data').get(CASE_FIELD_3_COLLECTION.id)).not.toBeNull();
+          expect((component.form.get('data').get(CASE_FIELD_3_COLLECTION.id) as FormArray).at(0)
+            .get('value.AddressLine1')).not.toBeNull();
+          expect((component.form.get('data').get(CASE_FIELD_3_COLLECTION.id) as FormArray).at(0)
+            .get('value.AddressLine1').value).toBeNull();
         });
       });
     });