+++ RELEASE-NOTES.md
@@ -2,6 +2,9 @@
 ### Version 2.64.50-case-share-hot-fix
 * correct and cherry-pick the right commits
 
+### Version 2.64.49-retain-hidden-value-complex-and-collection-types-fix
+**EUI-2739** Fix "retain hidden fields" functionality for all field types, including Complex and Collection types
+
 ### Version 2.64.48-case-share-hot-fix
 * determine whether a case sharable by supplementary_data